Add files via upload
Validation and Accuracy files includes txt files containing the exact and the estimated signatures for the chain network instances, txt files containing a summary of the experiments, an R script with the Hypothesis Test routine, and txt files with the corresponding P-values. The routines that creates the Chain networks and computes its exact survival signature are found in the BO_signature file.